backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-socketio>=5.10.0
